PORT-007: Renaming folder for curriculum vitae component
